threedimensional/requirements.txt
Adafruit-Blinka==6.16.0
adafruit-circuitpython-neopixel==6.2.3
adafruit-circuitpython-pixelbuf==1.1.1
Adafruit-PlatformDetect==3.18.0
Adafruit-PureIO==1.1.9
pkg-resources==0.0.0
pyftdi==0.53.3
pyserial==3.5
pyusb==1.2.1
rpi-ws281x==4.3.1
RPi.GPIO==0.7.0
sysv-ipc==1.1.0
